% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/diaClassification.R
\name{diaClassification}
\alias{diaClassification}
\title{Classic Dialektometrical ad-hoc classification

Applies certain ad-hoc measures used in Dialektometrics.

The available classification methods are Med MinMwMax and MedMw.}
\usage{
diaClassification(measure, method, n, id)
}
\arguments{
\item{measure}{Object coercible to data.frame containing the data.}

\item{method}{Classification method, see details.}

\item{n}{Number of groups to be made.}

\item{id}{vector of labels indicating which linguistical use as center. All will be used by default.}
}
\description{
Classic Dialektometrical ad-hoc classification

Applies certain ad-hoc measures used in Dialektometrics.

The available classification methods are Med MinMwMax and MedMw.
}
\examples{
data(dsample)
measure <- diaMeasure(dsample, location ~ question, 'answer', 'ird')
diaClassification(measure, 'med')
}

