% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/diaMeasure.R
\name{diaMeasure}
\alias{diaMeasure}
\title{Computes a Dialectometrical measure}
\usage{
diaMeasure(data, formula, value.var, measure = c("ird", "ipi"),
  binaryIndex = c("dice", "jaccard"), percentage = TRUE, subset)
}
\arguments{
\item{data}{object coercible to data.frame containing the data.}

\item{formula}{formula indicating which variables represent ids and variables, see details.}

\item{value.var}{character of length 1 indicating the variable containing.}

\item{measure}{character of length 1 indiciating the selected.}

\item{subset}{index indicating which subset of data to take.}
}
\description{
Computes a Dialectometrical measure
}
\examples{
data(dsample)
measure <- diaMeasure(simpleSample, gender + location ~ question, 'answer', 'ird')
print(measure)
}

