\name{diaMeasures-package}
\alias{diaMeasures-package}
\alias{diaMeasures}
\docType{package}
\title{
What the package does (short line).
}
\description{
More about what it does (maybe more than one line).
}
\details{
  \tabular{ll}{
    Package: \tab diaMeasures\cr
    Type: \tab Package\cr
    Version: \tab 1.0\cr
    Date: \tab 2015-08-14\cr
    License: \tab GPL-2\cr
  }
}
\author{
  Who wrote it Maintainer: Who to complain to <yourfault@somewhere.net>.
}
\references{
  no references.
}
\keyword{ package }
\seealso{
  what.
}
\examples{
## no examples
1*1
}
