% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/diaMeasure.R
\name{print.diaMeasure}
\alias{print.diaMeasure}
\title{Prints diaMeasure}
\usage{
\method{print}{diaMeasure}(x, lower = TRUE, justify = "none",
  digits = getOption("digits"))
}
\description{
Print lower or upper part of the coerced matrix.
}

