% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diaClassification.R
\name{diaClassification}
\alias{diaClassification}
\title{Classic Dialektometrical ad-hoc classification

Applies certain ad-hoc measures used in Dialektometrics. 

The available classification methods are Med MinMwMax and MedMw.}
\usage{
diaClassification(measure, method = c("Med", "MinMwMax", "MedMw"), n, ...)
}
\arguments{
\item{measure}{Object coercible to data.frame containing the data.}

\item{method}{Classification method, see details.}

\item{n}{Number of groups to be made.}

\item{...}{arguments passing to other method.}
}
\examples{

data(dsample)
measure <- diaMeasure(dsample, location ~ question, 'answer', 'ird')
diaClass <- diaClassification(measure, 'MinMwMax', 6)
diaClass2 <- diaClassification(measure, 'Med', 2, 'Itziar')

diaClassVector <- diaClassification(as.matrix(measure)[, 1], 'Med', 2)


}

