% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diaMeasure.R
\name{diaMeasure}
\alias{diaMeasure}
\title{Computes a Dialectometrical measure}
\usage{
diaMeasure(data, formula, value.var, measure = c("ird", "ipi", "levenshtein",
  "iri", "ipd"), binaryIndex = c("dice", "jaccard"), subset)
}
\arguments{
\item{data}{object coercible to data.frame containing the data.}

\item{formula}{formula indicating which variables represent ids and variables, see details.}

\item{value.var}{character of length 1 indicating the variable containing the values.}

\item{measure}{character of length 1 indiciating the selected.}

\item{binaryIndex}{character of length 1 indicating the binary index to use in multiple answers}

\item{subset}{index indicating which subset of data to take.}
}
\details{
Formula takes a expression of type 'var1 + var2 ~ var3' indication the index variables on
the LHS and reponse variables on the RHS.
}
\examples{

data(dsample)
measure <- diaMeasure(dsample, gender + location ~ question, 'answer', 'ird')
print(measure)

}

