% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diaMeasure.R
\name{print.diaMeasure}
\alias{print.diaMeasure}
\title{Prints diaMeasure}
\usage{
\method{print}{diaMeasure}(x, lower = TRUE, justify = "none",
  digits = getOption("digits"), ...)
}
\arguments{
\item{x}{object of clas diaMeasure.}

\item{lower}{logical of length one indicating if only the lower part should be passed.}

\item{justify}{argument that is passed to format.}

\item{digits}{digits to be shown.}

\item{...}{unused}
}
\description{
Print lower or upper part of the coerced matrix.
}

